Handle mapper (from dto to domain model)
