--- app/src/main/java/ru/practicum/android/diploma/data/mapper/VacancyResponseMapper.kt
@@ -1,39 +1,0 @@
-package ru.practicum.android.diploma.data.mapper
-
-import ru.practicum.android.diploma.data.dto.VacanciesSearchResponse
-import ru.practicum.android.diploma.data.dto.vacancy.ItemDto
-import ru.practicum.android.diploma.domain.models.Vacancy
-
-class VacancyResponseMapper {
-
-    fun map(response: VacanciesSearchResponse): List<Vacancy> {
-        return response.items.map { vacancy -> createVacancy(vacancy) }
-    }
-
-    private fun createVacancy(vacancy: ItemDto) = Vacancy(
-        id = vacancy.id,
-        imageUrl = vacancy.employer.logoUrls?.original,
-        name = vacancy.name,
-        city = getValueOrDefault(vacancy.area?.name, EMPTY_PARAM_VALUE),
-        salaryFrom = getValueOrDefault(vacancy.salary?.from?.toString(), EMPTY_PARAM_VALUE),
-        salaryTo = getValueOrDefault(vacancy.salary?.to?.toString(), EMPTY_PARAM_VALUE),
-        currency = getValueOrDefault(vacancy.salary?.currency, EMPTY_PARAM_VALUE),
-        employerName = vacancy.employer.name,
-        experience = getValueOrDefault(vacancy.experience?.name, EMPTY_PARAM_VALUE),
-        employmentName = getValueOrDefault(vacancy.employment?.name, EMPTY_PARAM_VALUE),
-        schedule = getValueOrDefault(vacancy.schedule?.name, EMPTY_PARAM_VALUE),
-        descriptionResponsibility = getValueOrDefault(vacancy.snippet.responsibility, EMPTY_PARAM_VALUE),
-        descriptionRequirement = getValueOrDefault(vacancy.snippet.requirement, EMPTY_PARAM_VALUE),
-        descriptionConditions = "???", // Может быть заполнено позже
-        descriptionSkills = "???" // Может быть заполнено позже
-    )
-
-    private fun <T> getValueOrDefault(
-        value: T?,
-        defaultValue: String
-    ): String = value?.toString() ?: defaultValue
-
-    companion object {
-        const val EMPTY_PARAM_VALUE = "Не указано"
-    }
-}
